requirements.txt
databricks-api==0.7.0
fastapi==0.75.0
python-dotenv==0.19.2
uvicorn==0.17.6
